requirements.txt
pandas 
numpy
matplotlib
seaborn 
scikit-learn
scipy
flask
pyaml
dill
mlflow
dvc
pickle
urllib


- e.